Split HollowNucleus to a separate file
