+++ src/CMakeLists.txt
@@ -16,6 +16,7 @@ general/PointNucleus.cpp
 general/RadialPotential.cpp
 general/GaussianNucleus.cpp
 general/SphericalNucleus.cpp
+general/HollowNucleus.cpp
 )
 target_link_libraries(helfem-common PUBLIC helfem)
 target_include_directories(helfem-common PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/../libhelfem/src/")
