+++ src/general/model_potential.cpp
@@ -26,28 +26,6 @@ namespace helfem {
       throw std::logic_error("Unrecognized model\n");
     }
 
-    HollowNucleus::HollowNucleus(int Z_, double R_) : Z(Z_), R(R_) {
-    }
-
-    HollowNucleus::~HollowNucleus() {
-    }
-
-    double HollowNucleus::V(double r) const {
-      if(r>=R) {
-        return -Z/r;
-      } else {
-        return -Z/R;
-      }
-    }
-
-    double HollowNucleus::get_R() const {
-      return R;
-    }
-
-    void HollowNucleus::set_R(double R_) {
-      R=R_;
-    }
-
     TFAtom::TFAtom(int Z_) : Z(Z_) {
     }
